rename compound_expression.* -> compound.*
