+++ src/semantics/readylog/compound.h
@@ -15,40 +15,16 @@
  * along with golog++.  If not, see <https://www.gnu.org/licenses/>.
 **************************************************************************/
 
-#ifndef GOLOGPP_COMPOUND_EXPRESSION_H_
-#define GOLOGPP_COMPOUND_EXPRESSION_H_
+#ifndef READYLOG_COMPOUND_EXPRESSION_H_
+#define READYLOG_COMPOUND_EXPRESSION_H_
 
-#include "language.h"
-#include "gologpp.h"
-#include "expressions.h"
-#include "scope.h"
-#include "types.h"
+#include "semantics.h"
 
-#include <unordered_map>
+#include <model/compound.h>
 
 namespace gologpp {
 
 
-class CompoundExpression
-: public Expression
-, public LanguageElement<CompoundExpression>
-, public NoScopeOwner
-{
-public:
-	CompoundExpression(const Type &type, const vector<fusion_wtf_vector<string, Expression *>> &entries);
+}
 
-	const Expression &entry(const string &key) const;
-
-	virtual void attach_semantics(SemanticsFactory &) override;
-	virtual string to_string(const string &pfx) const override;
-	const CompoundType &compound_type() const;
-
-private:
-	std::unordered_map<string, unique_ptr<Expression>> entries_;
-};
-
-
-
-} // namespace gologpp
-
-#endif // GOLOGPP_COMPOUND_EXPRESSION_H_
+#endif // READYLOG_COMPOUND_EXPRESSION_H_
